feat: add exceptions

+++ main-service/src/main/java/ru/practicum/categorie/CategoryServiceImpl.java
@@ -1,5 +1,6 @@
 package ru.practicum.categorie;
 
+import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Sort;
@@ -8,15 +9,20 @@ import ru.practicum.categorie.model.CategoryEntity;
 import ru.practicum.categorie.model.CategoryMapper;
 import ru.practicum.categorie.model.dto.CategoryDto;
 import ru.practicum.categorie.model.dto.NewCategoryDto;
+import ru.practicum.event.EventRepository;
+import ru.practicum.exception.ConstraintConflictException;
 import ru.practicum.exception.EntityNotFoundException;
 import ru.practicum.pagination.PaginationHelper;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
     private final CategoryMapper mapper;
 
     private CategoryEntity categoryExistCheck(Long catId) {
@@ -24,8 +30,16 @@ public class CategoryServiceImpl implements CategoryService {
                 new EntityNotFoundException("Category not found"));
     }
 
+    private void nameAlreadyExistCheck(String newName, Long categoryId) {
+        Optional<CategoryEntity> category = categoryRepository.findFirstByName(newName);
+        if(category.isPresent() && !Objects.equals(category.get().getId(), categoryId)) {
+            throw new ConstraintConflictException("Such categorie name already exsists .");
+        }
+    }
+
     @Override
     public CategoryDto create(NewCategoryDto newCategoryDto) {
+        nameAlreadyExistCheck(newCategoryDto.getName(), 0L);
         CategoryEntity entity = categoryRepository.save(mapper.toCategoryEntity(newCategoryDto));
         return mapper.toCategoryDto(entity);
     }
@@ -44,6 +58,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public CategoryDto update(Long catId, NewCategoryDto newCategoryDto) {
         CategoryEntity entity = categoryExistCheck(catId);
+        nameAlreadyExistCheck(newCategoryDto.getName(), entity.getId());
         mapper.update(newCategoryDto, entity);
         CategoryEntity updatedEntity = categoryRepository.save(entity);
         return mapper.toCategoryDto(updatedEntity);
@@ -51,6 +66,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Long catId) {
+        if(eventRepository.findFirstByCategoryId(catId).isPresent()) {
+            throw new ConstraintConflictException("With this category associated some events .");
+        }
         categoryRepository.deleteById(catId);
     }
